[MIG] website_sale_tax_toggle: Migration to 16.0
