+++ website_sale_tax_toggle/__manifest__.py
@@ -3,7 +3,7 @@
 {
     "name": "Website Sale Tax Toggle",
     "summary": "Allow display price in Shop with or without taxes",
-    "version": "15.0.1.0.1",
+    "version": "16.0.1.0.0",
     "development_status": "Beta",
     "category": "Website",
     "website": "https://github.com/OCA/e-commerce",
@@ -15,11 +15,11 @@
     "data": ["views/templates.xml"],
     "assets": {
         "web.assets_frontend": [
-            "/website_sale_tax_toggle/static/src/js/website_sale_tax_toggle.js",
+            "/website_sale_tax_toggle/static/src/js/website_sale_tax_toggle.esm.js",
             "/website_sale_tax_toggle/static/src/scss/website_sale_tax_toggle.scss",
         ],
         "web.assets_tests": [
-            "/website_sale_tax_toggle/static/src/js/website_sale_tax_toggle_tour.js"
+            "/website_sale_tax_toggle/static/src/js/website_sale_tax_toggle_tour.esm.js"
         ],
     },
 }
