+++ website_sale_tax_toggle/tests/test_website_sale_tax_toggle.py
@@ -1,21 +1,16 @@
 # Copyright 2020 Tecnativa - Sergio Teruel
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+from odoo.tests import tagged
 from odoo.tests.common import HttpCase
 
+from odoo.addons.base.tests.common import DISABLED_MAIL_CONTEXT
 
+
+@tagged("post_install", "-at_install")
 class WebsiteSaleTaxesToggleHttpCase(HttpCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        # Remove this variable in v16 and put instead:
-        # from odoo.addons.base.tests.common import DISABLED_MAIL_CONTEXT
-        DISABLED_MAIL_CONTEXT = {
-            "tracking_disable": True,
-            "mail_create_nolog": True,
-            "mail_create_nosubscribe": True,
-            "mail_notrack": True,
-            "no_reset_password": True,
-        }
         cls.env = cls.env(context=dict(cls.env.context, **DISABLED_MAIL_CONTEXT))
         # Get company for Mitchel Admin user
         cls.user_admin = cls.env.ref("base.user_admin")
@@ -47,13 +42,8 @@ class WebsiteSaleTaxesToggleHttpCase(HttpCase):
 
     def test_ui_website(self):
         """Test frontend tour."""
-        tour = (
-            "odoo.__DEBUG__.services['web_tour.tour']",
-            "website_sale_tax_toggle",
-        )
-        self.browser_js(
+        self.start_tour(
             url_path="/shop",
-            code="%s.run('%s')" % tour,
-            ready="%s.tours['%s'].ready" % tour,
+            tour_name="website_sale_tax_toggle",
             login="admin",
         )
